requirements.txt
# Core RAG and LangChain dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-deepseek>=0.1.0
langchain-chroma>=0.1.0

# RAGAS for evaluation
ragas>=0.1.0

# Vector database
chromadb>=0.4.0

# Data processing
pandas>=2.0.0
datasets>=2.14.0
numpy>=1.24.0

# PDF processing
pypdf>=3.15.0
unstructured[pdf]>=0.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing framework
pytest>=7.0.0
pytest-cov>=4.0.0

# Machine Learning
scikit-learn>=1.3.0

# Schema validation
jsonschema>=4.17.0

# Environment and utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: For better PDF processing
libmagic>=0.4.0